+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Release workflow tasks & github actions integration
 ### Fixed
   - Ensure all methods (including inherited ones) are considered when calculating signatures for a class or module.
+  - Fix calculation of total / filtered examples for rspec integration
 
 ## 0.1.0.pre.3
 ### Added